- Updated Smkinven to output VMT and SPEED after PMC when PMC is calculated with a formula.
- Corrected Smkinven to fix bug introduced in last release that doesn't allow
EMS-95 any hour-specific input files to be matched properly to the
inventory.  Changed TSCC assignment from file to be specific to the format of
the inventory and day- or hour-specific files.

- Updates to Smkinven to support hour-specific stack and plume rise information
(specifically done to allow external plume rise calculations for wildfires).

- Updates to Smkinven to correct bug that caused IDA and mobile EMS-95 formats
to not have duplicates summed, even though the code reported that the emissions
were summed.

- Updates to permit ozone-season emissions to be used to fill in annual
emissions when user-selection option is set in Smkinven.

- Correct bug in Smkinven for mobile sources (that didn't seem to affect any
runs) for which a zero index was set in POLVLA and POLVAL.

- Added SCC to day-specific and hour-specific formats for matching with IDA
files.  Make sure that this is in the newest version of the doc.

- Updates to permit hour-specific layer fraction information in Smkinven,
Laypoint, and Smkmerge (for UAM-style outputs)