- Updates to Smkmerge to correct bug in units conversion in which the numerator in the output units did not get changed correctly for *mass* based merges that included biogenics.  E.G., prep of MAQSIP inputs in g/sec.
- Updates to Smkmerge openmrgout to prevent a usually innocuous problem in which
the syntax of the code that set the vertical layers could cause a run-time
warning in debug mode about the extent of the arrays being dissimilar. Changed
the syntax to have better code (although the compilers seemed to handle the
poor syntax fine).

- Fixed Smkmerge again to output whatever units are asked for (with the same
notation), provided that it recognizes the untis. It does this while still
allowing automatic g/hr for PM species (for example)

- Changed openmrgin (in SMkmerge) to not abort when base inventory years do not
match and no projection is being done.  NEED TO TEST THIS MORE FOR VARIOUS
CONFIGURATIONS.  Prevents error abort when, for example, 1999 area emissions are
merged with 2001 mobile emissions.

- Fixed bug in Smkmerge that caused error when trying to merge pollutants and
activities from the MOBL file with MGMAT (and no MTMP file involved).

- Updates to permit hour-specific layer fraction information in Smkinven,
Laypoint, and Smkmerge (for UAM-style outputs)